more tests and move SI record generation to separate util class
